prediction-service/requirements.txt
confluent-kafka
python-dotenv
scikit-learn
pandas